compat: adapt to transformers 5.x deprecations
- from_pretrained: pass dtype= instead of deprecated torch_dtype= (model
  loading and MIA reference model). Requires transformers>=4.56.
- Remove deprecated TrainingArguments.logging_dir (removed in v5.2);
  TensorBoard now defaults under output_dir/runs.
Note: the YAML 'torch_dtype' key is the project's own config consumed by
get_dtype() and is intentionally left unchanged.

+++ src/evals/metrics/mia/__init__.py
@@ -87,7 +87,7 @@ def mia_reference(model, **kwargs):
     logger.info(f"Loading reference model from {kwargs['reference_model_path']}")
     reference_model = AutoModelForCausalLM.from_pretrained(
         kwargs["reference_model_path"],
-        torch_dtype=model.dtype,
+        dtype=model.dtype,  # transformers>=4.56 renamed `torch_dtype` -> `dtype`
         device_map={"": model.device},
     )
     return mia_auc(
